+++ public/css/main.css
@@ -142,6 +142,9 @@
 .plist_box_dj{
     display: flex;
     overflow: auto;
+    
+
+
     -ms-overflow-style: none; 
     /* 인터넷 익스플로러 */
     scrollbar-width: none; 
@@ -161,6 +164,17 @@
     width: 330px;
     height: 200px;
 }
+
+.plist_box_dj video{
+    width: 330px;
+    height: 200px;
+}
+
+/* .plist_box_dj div{
+    width: 330px;
+    height: 200px;
+    background-color: white;
+} */
 
 .relist_box_dj{
     display: flex;
